Big commit: making parsing more flexible
Big changes in homerFile and homerFrame classes. The code is overall simpler, but also way more flexible.
Now the input file is parsed as 'command' 'value', whatever 'value' is
(in particular whatever the number of fields it contains). It will allow
to implement flexible input format by having arbitrary column numbers.

There are some regressions, and there is so far no code to deal with exceptions in file reading.